SymbolicNumericIntegration.jl

SymbolicNumericIntegration.jl is a hybrid symbolic/numerical integration package that works on the Julia Symbolics expressions.

SymbolicNumericIntegration.jl uses a randomized algorithm based on a hybrid of the method of undetermined coefficients and sparse regression and can solve a large subset of basic standard integrals (polynomials, exponential/logarithmic, trigonometric and hyperbolic, inverse trigonometric and hyperbolic, rational and square root). The symbolic part of the algorithm is similar (but not identical) to Risch-Bronstein's poor man's integrator and generates a list of ansatzes (candidate terms). The numerical part uses sparse regression adopted from the Sparse identification of nonlinear dynamics (SINDy) algorithm to prune down the ansatzes and find the corresponding coefficients. The basis of how it works and the theory of integration using the Symbolic-Numeric methods refer to Basis of Symbolic-Numeric Integration.

hyint is the python counterpart of SymbolicNumericIntegration.jl that works with sympy expressions.

Originally, SymbolicNumericIntegration.jl expected only univariate expression with constant real or complex coefficients as input. As of version 1.2, integrate function accepts symbolic constants for a subset of solvable integrals.

integrate returns a tuple with three values. The first one is the solved integral, the second one is the sum of the unsolved terms, and the third value is the residual error. If integrate is successful, the unsolved portion is reported as 0. If we pass detailed=false to integrate', the output is simplified to only the resulting integrals. In this case,nothing` is returned if the integral cannot be solved. Note that the simplified output will become the default in a future version.

Installation

To install SymbolicNumericIntegration.jl, use the Julia package manager:

using Pkg
Pkg.add("SymbolicNumericIntegration")

Examples:

using Symbolics
using SymbolicNumericIntegration

@variables x a b

# if `detailed = true` (default), the output is a tuple of (solution, unsolved portion, err)

integrate(3x^3 + 2x - 5)
(x^2 + (3 // 4) * (x^4) - (5x), 0, 0)

integrate((5 + 2x)^-1)
((1 // 2) * log((5 // 2) + x), 0, 0.0)

# `detailed = false` simplifies the output to just the resulting integral

integrate(x^2 / (16 + x^2); detailed = false)
x + 4atan((-1 // 4) * x)

integrate(x^2 * log(x); detailed = false)
(1 // 3) * (x^3) * log(x) - (1 // 9) * (x^3)

integrate(sec(x) * tan(x); detailed = false)
sec(x)

# Symbolic integration. Here, a is a symbolic constant; therefore, we need 
# to explicitly define the independent variable (say, x). Also, we set
# `symbolic = true` to force using the symbolic solver

integrate(sin(a * x), x; detailed = false, symbolic = true)
(-cos(a * x)) / a

integrate(x^2 * cos(a * x), x; detailed = false, symbolic = true)
((a^2) * (x^2) * sin(a * x) + 2.0a * x * cos(a * x) - 2.0sin(a * x)) / (a^3)

integrate(log(log(a * x)) / x, x; detailed = false, symbolic = true)
log(a * x) * log(log(a * x)) - log(a * x)

# multiple symbolic constants

integrate(cosh(a * x) * exp(b * x), x; detailed = false, symbolic = true)
(a * sinh(a * x) * exp(b * x) - b * cosh(a * x) * exp(b * x)) / (a^2 - (b^2))

# definite integration, passing a tuple of (x, lower bound, higher bound) in the 
# second argument

integrate(x * sin(a * x), (x, 0, 1); symbolic = true, detailed = false)
(sin(a) - a * cos(a)) / (a^2)

SymbolicNumericIntegration.jl exports some special integral functions (defined over Complex numbers) and uses them in solving integrals:

Ei: exponential integral (defined as ∫ exp(x) / x dx)

Si: sine integral (defined as ∫ sin(x) / x dx)

Ci: cosine integral (defined as ∫ cos(x) / x dx)

Li: logarithmic integral (defined as ∫ 1 / log(x) dx)

For example:

integrate(exp(x + 1) / (x + 1))
(SymbolicNumericIntegration.Ei(1 + x), 0, 0.0)

integrate(x * cos(a * x^2 - 1) / (a * x^2 - 1), x; detailed = false, symbolic = true)
((1 // 2) * SymbolicNumericIntegration.Ci(a * (x^2) - 1)) / a

integrate(1 / (x * log(log(x))), x; detailed = false, symbolic = true)
SymbolicNumericIntegration.Li(log(x))

SymbolicNumericIntegration.integrate — Method

integrate(eq, x; kwargs...)

is the main entry point to integrate a univariate expression eq with respect to `x' (optional).

julia> using Symbolics, SymbolicNumericIntegration

julia> @variables x a

julia> integrate(x * sin(2x))
((1//4)*sin(2x) - (1//2)*x*cos(2x), 0, 0)

julia> integrate(x * sin(a * x), x; symbolic = true, detailed = false)
(sin(a*x) - a*x*cos(a*x)) / (a^2)

julia> integrate(x * sin(a * x), (x, 0, 1); symbolic = true, detailed = false)
(sin(a) - a*cos(a)) / (a^2)

Arguments:

eq: a univariate expression

x: independent variable (optional if eq is univariate) or a tuple of (independent variable, lower bound, upper bound) for definite integration.

Keyword Arguments:

abstol (default: 1e-6): the desired tolerance

num_steps (default: 2): the number of different steps with expanding basis to be tried

num_trials (default: 10): the number of trials in each step (no changes to the basis)

show_basis (default: false): if true, the basis (list of candidate terms) is printed

bypass (default: false): if true do not integrate terms separately but consider all at once

symbolic (default: false): try symbolic integration first (will be forced if eq has symbolic constants)

max_basis (default: 100): the maximum number of candidate terms to consider

verbose (default: false): print a detailed report

complex_plane (default: true): generate random test points on the complex plane (if false, the points will be on real axis)

radius (default: 1.0): the radius of the disk in the complex plane to generate random test points

opt (default: STLSQ(exp.(-10:1:0))): the sparse regression optimizer (from DataDrivenSparse)

homotopy (default: true): deprecated, will be removed in a future version

use_optim (default: false): deprecated, will be removed in a future version

detailed (default: true): (solved, unsolved, err) output format. If detailed=false, only the final integral is returned.

Returns a tuple of (solved, unsolved, err) if detailed == true, where

solved: the solved integral
unsolved: the residual unsolved portion of the input
err: the numerical error in reaching the solution

Returns the resulting integral or nothing if detailed == false

Testing

test/runtests.jl contains a test suite of 170 easy to moderate test integrals (can be run by calling test_integrals). Currently, SymbolicNumericIntegration.jl solves more than 95% of its test suite.

Additionally, 12 test suites from the Rule-based Integrator (Rubi) are included in the /test directory. For example, we can test the first one as below. Axiom refers to the format of the test files)

using SymbolicNumericIntegration
include("test/axiom.jl")  # note, you may need to use the correct path

L = convert_axiom(:Apostle)   # can also use L = convert_axiom(1)  
test_axiom(L, false; bypass = false, verbose = false, homotopy = true)

The test suites description based on the header of the files in the Rubi directory are

name	id	comment

:Apostle	1	Tom M. Apostol - Calculus, Volume I, Second Edition (1967)

:Bondarenko	2	Vladimir Bondarenko Integration Problems

:Bronstein	3	Manuel Bronstein - Symbolic Integration Tutorial (1998)

:Charlwood	4	Kevin Charlwood - Integration on Computer Algebra Systems (2008)

:Hearn	5	Anthony Hearn - Reduce Integration Test Package

:Hebisch	6	Waldek Hebisch - email May 2013

:Jeffrey	7	David Jeffrey - Rectifying Transformations for Trig Integration (1997)

:Moses	8	Joel Moses - Symbolic Integration Ph.D. Thesis (1967)

:Stewart	9	James Stewart - Calculus (1987)

:Timofeev	10	A. F. Timofeev - Integration of Functions (1948)

:Welz	11	Martin Welz - posts on Sci.Math.Symbolic

:Wester	12	Michael Wester

Citation

If you use SymbolicNumericIntegration.jl, please cite Symbolic-Numeric Integration of Univariate Expressions based on Sparse Regression:

@article{Iravanian2022,   
   author = {Shahriar Iravanian and Carl Julius Martensen and Alessandro Cheli and Shashi Gowda and Anand Jain and Julia Computing and Yingbo Ma and Chris Rackauckas},
   doi = {10.48550/arxiv.2201.12468},
   month = {1},
   title = {Symbolic-Numeric Integration of Univariate Expressions based on Sparse Regression},
   url = {https://arxiv.org/abs/2201.12468v2},
   year = {2022},
}

Contributing

Please refer to the SciML ColPrac: Contributor's Guide on Collaborative Practices for Community Packages for guidance on PRs, issues, and other matters relating to contributing to SciML.

See the SciML Style Guide for common coding practices and other style decisions.

There are a few community forums:

The #diffeq-bridged and #sciml-bridged channels in the Julia Slack

The #diffeq-bridged and #sciml-bridged channels in the Julia Zulip

On the Julia Discourse forums

See also SciML Community page

Reproducibility

The documentation of this SciML package was built using these direct dependencies,

Status `~/_work/SymbolicNumericIntegration.jl/SymbolicNumericIntegration.jl/docs/Project.toml`
  [e30172f5] Documenter v1.17.0
  [78aadeae] SymbolicNumericIntegration v1.11.1 `~/_work/SymbolicNumericIntegration.jl/SymbolicNumericIntegration.jl`

and using this machine and Julia version.

Julia Version 1.12.6
Commit 15346901f00 (2026-04-09 19:20 UTC)
Build Info:
  Official https://julialang.org release
Platform Info:
  OS: Linux (x86_64-linux-gnu)
  CPU: 128 × AMD EPYC 7513 32-Core Processor
  WORD_SIZE: 64
  LLVM: libLLVM-18.1.7 (ORCJIT, znver3)
  GC: Built with stock GC
Threads: 1 default, 1 interactive, 1 GC (on 4 virtual cores)
Environment:
  JULIA_CPU_THREADS = 4
  JULIA_NUM_PRECOMPILE_TASKS = 4

A more complete overview of all dependencies and their versions is also provided.

Status `~/_work/SymbolicNumericIntegration.jl/SymbolicNumericIntegration.jl/docs/Manifest.toml`
  [47edcb42] ADTypes v1.22.2
  [14f7f29c] AMD v0.5.3
  [a4c015fc] ANSIColoredPrinters v0.0.1
  [6e696c72] AbstractPlutoDingetjes v1.4.0
  [1520ce14] AbstractTrees v0.4.5
  [7d9f7c33] Accessors v0.1.45
  [79e6a3ab] Adapt v4.7.0
  [66dad0bd] AliasTables v1.1.3
  [ec485272] ArnoldiMethod v0.4.0
  [4fba245c] ArrayInterface v7.27.0
  [4c555306] ArrayLayouts v1.12.2
  [a9b6321e] Atomix v1.1.3
  [aae01518] BandedMatrices v1.11.0
  [e2ed5e7c] Bijections v0.2.2
  [b2a6c25c] BinaryHeaps v1.0.3
  [caf10ac8] BipartiteGraphs v0.1.9
  [8e7c35d0] BlockArrays v1.9.5
  [70df07ce] BracketingNonlinearSolve v1.12.2
  [d360d2e6] ChainRulesCore v1.26.1
  [da1fd8a2] CodeTracking v3.0.2
  [944b1d66] CodecZlib v0.7.8
⌅ [861a8166] Combinatorics v1.0.2
  [38540f10] CommonSolve v0.2.11
  [bbf7d656] CommonSubexpressions v0.3.1
  [f70d9fcc] CommonWorldInvalidations v1.1.1
  [34da2185] Compat v4.18.1
  [b152e2b5] CompositeTypes v0.1.4
  [a33af91c] CompositionsBase v0.1.2
  [2569d6c7] ConcreteStructs v0.2.6
  [187b0558] ConstructionBase v1.6.0
  [a8cc5b0e] Crayons v4.1.1
  [9a962f9c] DataAPI v1.16.0
  [2445eb08] DataDrivenDiffEq v1.15.2
  [5b588203] DataDrivenSparse v0.1.5
  [82cc6244] DataInterpolations v9.0.2
  [864edb3b] DataStructures v0.19.5
  [e2d170a0] DataValueInterfaces v1.0.0
  [8bb1440f] DelimitedFiles v1.9.1
  [2b5f629d] DiffEqBase v7.6.1
  [459566f4] DiffEqCallbacks v4.18.3
  [163ba53b] DiffResults v1.1.0
  [b552c78f] DiffRules v1.16.0
  [a0c0ee7d] DifferentiationInterface v0.7.20
  [ffbed154] DocStringExtensions v0.9.5
  [e30172f5] Documenter v1.17.0
  [5b8099bc] DomainSets v0.8.1
  [7c1d4256] DynamicPolynomials v0.6.6
  [4e289a0a] EnumX v1.0.7
  [f151be2c] EnzymeCore v0.8.21
  [e2ba6199] ExprTools v0.1.10
  [55351af7] ExproniconLite v0.10.14
  [7034ab61] FastBroadcast v1.3.4
  [9aa1b823] FastClosures v0.3.2
  [a4df4552] FastPower v1.3.4
  [1a297f60] FillArrays v1.16.0
  [64ca27bc] FindFirstFunctions v3.0.3
  [6a86dc24] FiniteDiff v2.32.0
  [f6369f11] ForwardDiff v1.4.1
  [a85aefff] FunctionMaps v0.1.2
  [069b7b12] FunctionWrappers v1.1.3
  [77dc65aa] FunctionWrappersWrappers v1.10.1
  [46192b85] GPUArraysCore v0.2.0
  [d7ba0133] Git v1.5.0
  [86223c79] Graphs v1.14.0
  [076d061b] HashArrayMappedTries v0.2.0
  [b5f81e59] IOCapture v1.0.0
  [3263718b] ImplicitDiscreteSolve v2.1.2
  [d25df0c9] Inflate v0.1.5
  [18e54dd8] IntegerMathUtils v0.1.3
  [8197267c] IntervalSets v0.7.14
  [3587e190] InverseFunctions v0.1.17
  [92d709cd] IrrationalConstants v0.2.6
  [82899510] IteratorInterfaceExtensions v1.0.0
  [692b3bcd] JLLWrappers v1.8.0
  [682c06a0] JSON v1.6.1
  [ae98c720] Jieko v0.2.1
  [ccbc3e58] JumpProcesses v9.29.1
  [63c18a36] KernelAbstractions v0.9.42
  [ba0b0d4f] Krylov v0.10.8
  [b964fa9f] LaTeXStrings v1.4.0
  [0e77f7df] LazilyInitializedFields v1.3.0
  [87fe0de2] LineSearch v0.1.12
⌅ [7ed4a6bd] LinearSolve v4.3.0
  [2ab3a3ac] LogExpFunctions v1.0.1
  [e6f89c97] LoggingExtras v1.2.0
  [c2834f40] MLCore v1.1.0
  [f1d291b0] MLUtils v0.4.12
  [1914dd2f] MacroTools v0.5.16
  [d0879d2d] MarkdownAST v0.1.3
  [bb5d69b7] MaybeInplace v0.1.6
  [e1d29d7a] Missings v1.2.0
  [961ee093] ModelingToolkit v11.31.1
  [7771a370] ModelingToolkitBase v1.51.0
  [6bb917b9] ModelingToolkitTearing v1.18.2
  [2e0e35c7] Moshi v0.3.10
  [46d2c3a1] MuladdMacro v0.2.6
  [102ac46a] MultivariatePolynomials v0.5.19
  [d8a4904e] MutableArithmetics v1.8.0
  [872c559c] NNlib v0.9.38
  [77ba4419] NaNMath v1.1.4
  [be0214bd] NonlinearSolveBase v2.33.0
  [5959db7a] NonlinearSolveFirstOrder v2.1.3
  [6fe1bfb0] OffsetArrays v1.17.0
⌅ [bac558e1] OrderedCollections v1.8.2
  [bbf590c4] OrdinaryDiffEqCore v4.6.0
  [d96e819e] Parameters v0.13.1
  [69de0a69] Parsers v2.8.6
  [e409e4f3] PoissonRandom v0.4.12
  [d236fae5] PreallocationTools v1.3.0
  [aea7be01] PrecompileTools v1.3.4
  [21216c6a] Preferences v1.5.2
  [08abe8d2] PrettyTables v3.4.0
  [27ebfcd6] Primes v0.5.7
  [43287f4e] PtrArrays v1.4.0
  [0c0d3e7f] PureKLU v1.1.1
  [1fd47b50] QuadGK v2.11.3
  [988b38a3] ReadOnlyArrays v0.2.0
  [795d4caa] ReadOnlyDicts v1.0.1
  [3cdcf5f2] RecipesBase v1.3.4
  [731186ca] RecursiveArrayTools v4.3.3
  [189a3867] Reexport v1.2.2
  [2792f1a3] RegistryInstances v0.1.0
  [ae029012] Requires v1.3.1
  [7e49a35a] RuntimeGeneratedFunctions v0.5.22
  [9dfe8606] SCCNonlinearSolve v1.13.2
  [0bca4576] SciMLBase v3.35.1
  [19f34311] SciMLJacobianOperators v0.1.14
  [a6db7da4] SciMLLogging v2.0.3
  [c0aeaf25] SciMLOperators v1.24.0
  [431bcebd] SciMLPublic v1.2.2
  [53ae85a6] SciMLStructures v1.10.3
  [7e506255] ScopedValues v1.6.2
  [efcf1570] Setfield v1.1.2
  [605ecd9f] ShowCases v0.1.0
  [727e6d20] SimpleNonlinearSolve v2.12.1
  [699a6c99] SimpleTraits v0.9.6
  [a2af1166] SortingAlgorithms v1.2.3
  [a57abbd0] SparseColumnPivotedQR v2.1.4
  [276daf66] SpecialFunctions v2.8.0
  [0c0c59c1] StarAlgebras v0.3.0
  [64909d44] StateSelection v1.11.0
  [90137ffa] StaticArrays v1.9.18
  [1e83bf80] StaticArraysCore v1.4.4
  [10745b16] Statistics v1.11.1
  [82ae8749] StatsAPI v1.8.0
  [2913bbd2] StatsBase v0.34.12
  [892a3eda] StringManipulation v0.4.4
  [ec057cc2] StructUtils v2.8.2
  [2efcf032] SymbolicIndexingInterface v0.3.51
  [19f23fe9] SymbolicLimits v1.1.3
  [78aadeae] SymbolicNumericIntegration v1.11.1 `~/_work/SymbolicNumericIntegration.jl/SymbolicNumericIntegration.jl`
  [d1185830] SymbolicUtils v4.40.0
  [0c5d862f] Symbolics v7.31.0
  [3783bdb8] TableTraits v1.0.1
  [bd369af6] Tables v1.13.0
  [ed4db957] TaskLocalValues v0.1.3
  [8ea1fca8] TermInterface v2.0.0
  [a759f4b9] TimerOutputs v0.5.29
  [3bb67fe8] TranscodingStreams v0.11.3
  [781d530d] TruncatedStacktraces v1.4.0
  [3a884ed6] UnPack v1.0.2
  [013be700] UnsafeAtomics v0.3.1
  [d30d5f5c] WeakCacheSets v0.1.0
  [2e619515] Expat_jll v2.8.2+0
  [020c3dae] Git_LFS_jll v3.7.1+0
  [f8c6e375] Git_jll v2.54.0+0
  [1d5cc7b8] IntelOpenMP_jll v2025.2.0+0
  [94ce4f54] Libiconv_jll v1.18.0+0
  [856f044c] MKL_jll v2025.2.0+0
  [9bd350c2] OpenSSH_jll v10.4.1+0
  [efe28fd5] OpenSpecFun_jll v0.5.6+0
  [1317d2d5] oneTBB_jll v2022.3.0+0
  [0dad84c5] ArgTools v1.1.2
  [56f22d72] Artifacts v1.11.0
  [2a0f44e3] Base64 v1.11.0
  [ade2ca70] Dates v1.11.0
  [8ba89e20] Distributed v1.11.0
  [f43a241f] Downloads v1.7.0
  [7b1f6079] FileWatching v1.11.0
  [9fa8497b] Future v1.11.0
  [b77e0a4c] InteractiveUtils v1.11.0
  [ac6e5ff7] JuliaSyntaxHighlighting v1.12.0
  [4af54fe1] LazyArtifacts v1.11.0
  [b27032c2] LibCURL v0.6.4
  [76f85450] LibGit2 v1.11.0
  [8f399da3] Libdl v1.11.0
  [37e2e46d] LinearAlgebra v1.12.0
  [56ddb016] Logging v1.11.0
  [d6f4376e] Markdown v1.11.0
  [a63ad114] Mmap v1.11.0
  [ca575930] NetworkOptions v1.3.0
  [44cfe95a] Pkg v1.12.1
  [de0858da] Printf v1.11.0
  [3fa0cd96] REPL v1.11.0
  [9a3f8284] Random v1.11.0
  [ea8e919c] SHA v0.7.0
  [9e88b42a] Serialization v1.11.0
  [6462fe0b] Sockets v1.11.0
  [2f01184e] SparseArrays v1.12.0
  [f489334b] StyledStrings v1.11.0
  [fa267f1f] TOML v1.0.3
  [a4e569a6] Tar v1.10.0
  [8dfed614] Test v1.11.0
  [cf7118a7] UUIDs v1.11.0
  [4ec0a83e] Unicode v1.11.0
  [e66e0078] CompilerSupportLibraries_jll v1.3.0+1
  [deac9b47] LibCURL_jll v8.15.0+0
  [e37daf67] LibGit2_jll v1.9.0+0
  [29816b5a] LibSSH2_jll v1.11.3+1
  [14a3606d] MozillaCACerts_jll v2025.11.4
  [4536629a] OpenBLAS_jll v0.3.29+0
  [05823500] OpenLibm_jll v0.8.7+0
  [458c3c95] OpenSSL_jll v3.5.4+0
  [efcefdf7] PCRE2_jll v10.44.0+1
  [bea87d4a] SuiteSparse_jll v7.8.3+2
  [83775a58] Zlib_jll v1.3.1+2
  [8e850b90] libblastrampoline_jll v5.15.0+0
  [8e850ede] nghttp2_jll v1.64.0+1
  [3f19e933] p7zip_jll v17.7.0+0
Info Packages marked with ⌅ have new versions available but compatibility constraints restrict them from upgrading. To see why use `status --outdated -m`

You can also download the manifest file and the project file.