SymbolicNumericIntegration.jl

SymbolicNumericIntegration.jl is a hybrid symbolic/numerical integration package that works on the Julia Symbolics expressions.

SymbolicNumericIntegration.jl uses a randomized algorithm based on a hybrid of the method of undetermined coefficients and sparse regression and can solve a large subset of basic standard integrals (polynomials, exponential/logarithmic, trigonometric and hyperbolic, inverse trigonometric and hyperbolic, rational and square root). The basis of how it works and the theory of integration using the Symbolic-Numeric methods refer to Basis of Symbolic-Numeric Integration.

Function integrate returns the integral of a univariate expression with constant real or complex coefficients. integrate returns a tuple with three values. The first one is the solved integral, the second one is the sum of the unsolved terms, and the third value is the residual error. If integrate is successful, the unsolved portion is reported as 0.

Installation

To install SymbolicNumericIntegration.jl, use the Julia package manager:

using Pkg
Pkg.add("SymbolicNumericIntegration")

Examples:

using Symbolics
using SymbolicNumericIntegration

@variables x

julia> integrate(3x^3 + 2x - 5)
(x^2 + (3//4)*(x^4) - (5x), 0, 0)

julia> integrate((5 + 2x)^-1)
((1//2)*log((5//2) + x), 0, 0.0)

julia> integrate(1 / (6 + x^2 - (5x)))
(log(x - 3) - log(x - 2), 0, 3.339372764128952e-16)

julia> integrate(1 / (x^2 - 16))
((1//8)*log(x - 4) - ((1//8)*log(4 + x)), 0, 1.546926788028958e-16)

julia> integrate(x^2 / (16 + x^2))
(x + 4atan((-1//4)*x), 0, 1.3318788420751984e-16)

julia> integrate(x^2 / sqrt(4 + x^2))
((1//2)*x*((4 + x^2)^0.5) - ((2//1)*log(x + sqrt(4 + x^2))), 0, 8.702422633074313e-17)

julia> integrate(x^2 * log(x))
((1//3)*log(x)*(x^3) - ((1//9)*(x^3)), 0, 0)

julia> integrate(x^2 * exp(x))
(2exp(x) + exp(x)*(x^2) - (2x*exp(x)), 0, 0)

julia> integrate(tan(2x))
((-1//2)*log(cos(2x)), 0, 0)

julia> integrate(sec(x) * tan(x))
(cos(x)^-1, 0, 0)

julia> integrate(cosh(2x) * exp(x))
((2//3)*exp(x)*sinh(2x) - ((1//3)*exp(x)*cosh(2x)), 0, 7.073930088880992e-8)

julia> integrate(cosh(x) * sin(x))
((1//2)*sin(x)*sinh(x) - ((1//2)*cos(x)*cosh(x)), 0, 4.8956233716268386e-17)

julia> integrate(cosh(2x) * sin(3x))
(0.153845sinh(2x)*sin(3x) - (0.23077cosh(2x)*cos(3x)), 0, 4.9807620877373405e-6)

julia> integrate(log(log(x)) * (x^-1))
(log(x)*log(log(x)) - log(x), 0, 0)

julia> integrate(exp(x^2))
(0, exp(x^2), Inf)    # as expected!

SymbolicNumericIntegration.jl exports some special integral functions (defined over Complex numbers) and uses them in solving integrals:

Ei: exponential integral (define as ∫ exp(x) / x dx)

Si: sine integral (define as ∫ sin(x) / x dx)

Ci: cosine integral (define as ∫ cos(x) / x dx)

Li: logarithmic integral (define as ∫ 1 / log(x) dx)

For examples:

julia> integrate(exp(x + 1) / (x + 1))
(SymbolicNumericIntegration.Ei(1 + x), 0, 1.1796119636642288e-16)

julia> integrate(x * cos(x^2 - 1) / (x^2 - 1))
((1//2)*SymbolicNumericIntegration.Ci(x^2 - 1), 0, 2.7755575615628914e-17)

julia> integrate(1 / (x*log(log(x))))
(SymbolicNumericIntegration.Li(log(x)), 0, 1.1102230246251565e-16)

SymbolicNumericIntegration.integrate — Method

integrate(eq, x; kwargs...)

is the main entry point to integrate a univariate expression eq with respect to `x' (optional).

integrate(x * sin(2x))

# output

((1//4)*sin(2x) - (1//2)*x*cos(2x), 0, 0)

Arguments:

eq: a univariate expression

x: the independent variable (optional)

Keyword Arguments:

abstol (default: 1e-6): the desired tolerance

num_steps (default: 2): the number of different steps with expanding basis to be tried

num_trials (default: 10): the number of trials in each step (no changes to the basis)

show_basis (default: false): if true, the basis (list of candidate terms) is printed

bypass (default: false): if true do not integrate terms separately but consider all at once

symbolic (default: false): try symbolic integration first

max_basis (default: 100): the maximum number of candidate terms to consider

verbose (default: false): print a detailed report

complex_plane (default: true): generate random test points on the complex plane (if false, the points will be on real axis)

radius (default: 1.0): the radius of the disk in the complex plane to generate random test points

opt (default: STLSQ(exp.(-10:1:0))): the sparse regression optimizer (from DataDrivenSparse)

homotopy (default: true): use the homotopy algorithm to generate the basis (deprecated, will be removed in a future version)

use_optim (default: false): use Optim.jl minimize function instead of the STLSQ algorithm (experimental)

Output:

solved: the solved integral

unsolved: the residual unsolved portion of the input

err: the numerical error in reaching the solution

Testing

test/runtests.jl contains a test suite of 170 easy to moderate test integrals (can be run by calling test_integrals). Currently, SymbolicNumericIntegration.jl solves more than 95% of its test suite.

Additionally, 12 test suites from the Rule-based Integrator (Rubi) are included in the /test directory. For example, we can test the first one as below. Axiom refers to the format of the test files)

using SymbolicNumericIntegration
include("test/axiom.jl")  # note, you may need to use the correct path

L = convert_axiom(:Apostle)   # can also use L = convert_axiom(1)  
test_axiom(L, false; bypass = false, verbose = false, homotopy = true)

The test suites description based on the header of the files in the Rubi directory are

name	id	comment

:Apostle	1	Tom M. Apostol - Calculus, Volume I, Second Edition (1967)

:Bondarenko	2	Vladimir Bondarenko Integration Problems

:Bronstein	3	Manuel Bronstein - Symbolic Integration Tutorial (1998)

:Charlwood	4	Kevin Charlwood - Integration on Computer Algebra Systems (2008)

:Hearn	5	Anthony Hearn - Reduce Integration Test Package

:Hebisch	6	Waldek Hebisch - email May 2013

:Jeffrey	7	David Jeffrey - Rectifying Transformations for Trig Integration (1997)

:Moses	8	Joel Moses - Symbolic Integration Ph.D. Thesis (1967)

:Stewart	9	James Stewart - Calculus (1987)

:Timofeev	10	A. F. Timofeev - Integration of Functions (1948)

:Welz	11	Martin Welz - posts on Sci.Math.Symbolic

:Webster	12	Michael Wester

Citation

If you use SymbolicNumericIntegration.jl, please cite Symbolic-Numeric Integration of Univariate Expressions based on Sparse Regression:

@article{Iravanian2022,   
   author = {Shahriar Iravanian and Carl Julius Martensen and Alessandro Cheli and Shashi Gowda and Anand Jain and Julia Computing and Yingbo Ma and Chris Rackauckas},
   doi = {10.48550/arxiv.2201.12468},
   month = {1},
   title = {Symbolic-Numeric Integration of Univariate Expressions based on Sparse Regression},
   url = {https://arxiv.org/abs/2201.12468v2},
   year = {2022},
}

Contributing

Please refer to the SciML ColPrac: Contributor's Guide on Collaborative Practices for Community Packages for guidance on PRs, issues, and other matters relating to contributing to SciML.

See the SciML Style Guide for common coding practices and other style decisions.

There are a few community forums:

The #diffeq-bridged and #sciml-bridged channels in the Julia Slack

The #diffeq-bridged and #sciml-bridged channels in the Julia Zulip

On the Julia Discourse forums

See also SciML Community page

Reproducibility

The documentation of this SciML package was built using these direct dependencies,

Status `~/work/SymbolicNumericIntegration.jl/SymbolicNumericIntegration.jl/docs/Project.toml`
  [e30172f5] Documenter v0.27.24
  [78aadeae] SymbolicNumericIntegration v1.1.0 `~/work/SymbolicNumericIntegration.jl/SymbolicNumericIntegration.jl`

and using this machine and Julia version.

Julia Version 1.9.0
Commit 8e630552924 (2023-05-07 11:25 UTC)
Platform Info:
  OS: Linux (x86_64-linux-gnu)
  CPU: 2 × Intel(R) Xeon(R) Platinum 8370C CPU @ 2.80GHz
  WORD_SIZE: 64
  LIBM: libopenlibm
  LLVM: libLLVM-14.0.6 (ORCJIT, icelake-server)
  Threads: 1 on 2 virtual cores

A more complete overview of all dependencies and their versions is also provided.

Status `~/work/SymbolicNumericIntegration.jl/SymbolicNumericIntegration.jl/docs/Manifest.toml`
  [a4c015fc] ANSIColoredPrinters v0.0.1
⌅ [c3fe647b] AbstractAlgebra v0.29.4
  [1520ce14] AbstractTrees v0.4.4
  [7d9f7c33] Accessors v0.1.30
  [79e6a3ab] Adapt v3.6.2
  [dce04be8] ArgCheck v2.3.0
  [ec485272] ArnoldiMethod v0.2.0
  [4fba245c] ArrayInterface v7.4.4
  [a9b6321e] Atomix v0.1.0
  [198e06fe] BangBang v0.3.38
  [9718e550] Baselet v0.1.1
  [e2ed5e7c] Bijections v0.1.4
  [62783981] BitTwiddlingConvenienceFunctions v0.1.5
  [fa961155] CEnum v0.4.2
  [2a0fbf3d] CPUSummary v0.2.2
  [00ebfdb7] CSTParser v3.3.6
  [49dc2e85] Calculus v0.5.1
  [d360d2e6] ChainRulesCore v1.16.0
  [fb6a15b2] CloseOpenIntervals v0.1.12
  [861a8166] Combinatorics v1.0.2
  [a80b9123] CommonMark v0.8.12
  [38540f10] CommonSolve v0.2.3
  [bbf7d656] CommonSubexpressions v0.3.0
  [34da2185] Compat v4.6.1
  [b152e2b5] CompositeTypes v0.1.3
  [a33af91c] CompositionsBase v0.1.2
  [187b0558] ConstructionBase v1.5.2
  [6add18c4] ContextVariablesX v0.1.3
  [adafc99b] CpuId v0.3.1
  [a8cc5b0e] Crayons v4.1.1
  [9a962f9c] DataAPI v1.15.0
  [2445eb08] DataDrivenDiffEq v1.2.0
  [5b588203] DataDrivenSparse v0.1.2
  [82cc6244] DataInterpolations v4.0.0
  [864edb3b] DataStructures v0.18.13
  [e2d170a0] DataValueInterfaces v1.0.0
  [244e2a9f] DefineSingletons v0.1.2
  [8bb1440f] DelimitedFiles v1.9.1
  [2b5f629d] DiffEqBase v6.124.0
  [459566f4] DiffEqCallbacks v2.26.1
  [163ba53b] DiffResults v1.1.0
  [b552c78f] DiffRules v1.13.0
  [b4f34e82] Distances v0.10.8
  [31c24e10] Distributions v0.25.92
  [ffbed154] DocStringExtensions v0.9.3
  [e30172f5] Documenter v0.27.24
  [5b8099bc] DomainSets v0.6.6
  [fa6b7ba4] DualNumbers v0.6.8
  [7c1d4256] DynamicPolynomials v0.4.6
  [4e289a0a] EnumX v1.0.4
  [e2ba6199] ExprTools v0.1.9
  [cc61a311] FLoops v0.2.1
  [b9860ae5] FLoopsBase v0.1.1
  [7034ab61] FastBroadcast v0.2.5
  [1a297f60] FillArrays v1.0.0
  [6a86dc24] FiniteDiff v2.20.0
  [9c68100b] FoldsThreads v0.1.1
  [59287772] Formatting v0.4.2
  [f6369f11] ForwardDiff v0.10.35
  [069b7b12] FunctionWrappers v1.1.3
  [77dc65aa] FunctionWrappersWrappers v0.1.3
  [46192b85] GPUArraysCore v0.1.4
  [c27321d9] Glob v1.3.1
  [86223c79] Graphs v1.8.0
  [0b43b601] Groebner v0.3.4
  [d5909c97] GroupsCore v0.4.0
  [34004b35] HypergeometricFunctions v0.3.16
  [b5f81e59] IOCapture v0.2.2
  [615f187c] IfElse v0.1.1
  [d25df0c9] Inflate v0.1.3
  [22cec73e] InitialValues v0.3.1
  [18e54dd8] IntegerMathUtils v0.1.0
  [8197267c] IntervalSets v0.7.4
  [3587e190] InverseFunctions v0.1.9
  [92d709cd] IrrationalConstants v0.2.2
  [82899510] IteratorInterfaceExtensions v1.0.0
  [692b3bcd] JLLWrappers v1.4.1
  [682c06a0] JSON v0.21.4
  [98e50ef6] JuliaFormatter v1.0.29
  [b14d175d] JuliaVariables v0.2.4
  [ccbc3e58] JumpProcesses v9.6.3
  [63c18a36] KernelAbstractions v0.9.4
  [929cbde3] LLVM v5.1.0
  [b964fa9f] LaTeXStrings v1.3.0
  [2ee39098] LabelledArrays v1.14.0
  [984bce1d] LambertW v0.4.6
  [23fbe1c1] Latexify v0.16.0
  [10f19ff3] LayoutPointers v0.1.14
  [50d2b5c4] Lazy v0.15.1
  [d3d80556] LineSearches v7.2.0
  [2ab3a3ac] LogExpFunctions v0.3.23
  [d8e11817] MLStyle v0.4.17
  [f1d291b0] MLUtils v0.4.2
  [1914dd2f] MacroTools v0.5.10
  [d125e4d3] ManualMemory v0.1.8
  [128add7d] MicroCollections v0.1.4
  [e1d29d7a] Missings v1.1.0
  [961ee093] ModelingToolkit v8.55.1
  [46d2c3a1] MuladdMacro v0.2.4
  [102ac46a] MultivariatePolynomials v0.4.7
  [d8a4904e] MutableArithmetics v1.3.0
  [d41bc354] NLSolversBase v7.8.3
  [2774e3e8] NLsolve v4.5.1
  [872c559c] NNlib v0.8.20
  [77ba4419] NaNMath v1.0.2
  [71a1bf82] NameResolution v0.1.5
  [429524aa] Optim v1.7.5
  [bac558e1] OrderedCollections v1.6.0
  [90014a1f] PDMats v0.11.17
  [d96e819e] Parameters v0.12.3
  [69de0a69] Parsers v2.5.9
  [e409e4f3] PoissonRandom v0.4.4
  [f517fe37] Polyester v0.7.3
  [1d0040c9] PolyesterWeave v0.2.1
  [85a6dd25] PositiveFactorizations v0.2.4
  [d236fae5] PreallocationTools v0.4.12
  [aea7be01] PrecompileTools v1.1.1
  [21216c6a] Preferences v1.4.0
  [8162dcfd] PrettyPrint v0.2.0
  [27ebfcd6] Primes v0.5.3
  [92933f4c] ProgressMeter v1.7.2
  [1fd47b50] QuadGK v2.8.2
  [fb686558] RandomExtensions v0.4.3
  [e6cf234a] RandomNumbers v1.5.3
  [3cdcf5f2] RecipesBase v1.3.4
  [731186ca] RecursiveArrayTools v2.38.5
  [189a3867] Reexport v1.2.2
  [ae029012] Requires v1.3.0
  [79098fc4] Rmath v0.7.1
  [7e49a35a] RuntimeGeneratedFunctions v0.5.7
  [94e857df] SIMDTypes v0.1.0
  [0bca4576] SciMLBase v1.92.0
  [c0aeaf25] SciMLOperators v0.2.8
  [efcf1570] Setfield v1.1.1
  [605ecd9f] ShowCases v0.1.0
  [727e6d20] SimpleNonlinearSolve v0.1.14
  [699a6c99] SimpleTraits v0.9.4
  [66db9d55] SnoopPrecompile v1.0.3
  [a2af1166] SortingAlgorithms v1.1.0
  [276daf66] SpecialFunctions v2.2.0
  [171d559e] SplittablesBase v0.1.15
  [aedffcd0] Static v0.8.7
  [0d7ed370] StaticArrayInterface v1.4.0
  [90137ffa] StaticArrays v1.5.25
  [1e83bf80] StaticArraysCore v1.4.0
  [82ae8749] StatsAPI v1.6.0
⌅ [2913bbd2] StatsBase v0.33.21
  [4c63d2b9] StatsFuns v1.3.0
  [7792a7ef] StrideArraysCore v0.4.14
  [2efcf032] SymbolicIndexingInterface v0.2.2
  [78aadeae] SymbolicNumericIntegration v1.1.0 `~/work/SymbolicNumericIntegration.jl/SymbolicNumericIntegration.jl`
  [d1185830] SymbolicUtils v1.0.5
  [0c5d862f] Symbolics v5.3.1
  [3783bdb8] TableTraits v1.0.1
  [bd369af6] Tables v1.10.1
  [8290d209] ThreadingUtilities v0.5.1
  [a759f4b9] TimerOutputs v0.5.23
  [0796e94c] Tokenize v0.5.25
  [28d57a85] Transducers v0.4.76
  [a2a6695c] TreeViews v0.3.0
  [410a4b4d] Tricks v0.1.7
  [781d530d] TruncatedStacktraces v1.3.0
  [5c2747f8] URIs v1.4.2
  [3a884ed6] UnPack v1.0.2
  [1986cc42] Unitful v1.14.0
  [a7c27f48] Unityper v0.1.2
  [013be700] UnsafeAtomics v0.2.1
  [d80eeb9a] UnsafeAtomicsLLVM v0.1.2
  [700de1a5] ZygoteRules v0.2.3
  [dad2f222] LLVMExtra_jll v0.0.21+0
  [efe28fd5] OpenSpecFun_jll v0.5.5+0
  [f50d1b31] Rmath_jll v0.4.0+0
  [0dad84c5] ArgTools v1.1.1
  [56f22d72] Artifacts
  [2a0f44e3] Base64
  [ade2ca70] Dates
  [8ba89e20] Distributed
  [f43a241f] Downloads v1.6.0
  [7b1f6079] FileWatching
  [9fa8497b] Future
  [b77e0a4c] InteractiveUtils
  [4af54fe1] LazyArtifacts
  [b27032c2] LibCURL v0.6.3
  [76f85450] LibGit2
  [8f399da3] Libdl
  [37e2e46d] LinearAlgebra
  [56ddb016] Logging
  [d6f4376e] Markdown
  [a63ad114] Mmap
  [ca575930] NetworkOptions v1.2.0
  [44cfe95a] Pkg v1.9.0
  [de0858da] Printf
  [3fa0cd96] REPL
  [9a3f8284] Random
  [ea8e919c] SHA v0.7.0
  [9e88b42a] Serialization
  [1a1011a3] SharedArrays
  [6462fe0b] Sockets
  [2f01184e] SparseArrays
  [10745b16] Statistics v1.9.0
  [4607b0f0] SuiteSparse
  [fa267f1f] TOML v1.0.3
  [a4e569a6] Tar v1.10.0
  [8dfed614] Test
  [cf7118a7] UUIDs
  [4ec0a83e] Unicode
  [e66e0078] CompilerSupportLibraries_jll v1.0.2+0
  [deac9b47] LibCURL_jll v7.84.0+0
  [29816b5a] LibSSH2_jll v1.10.2+0
  [c8ffd9c3] MbedTLS_jll v2.28.2+0
  [14a3606d] MozillaCACerts_jll v2022.10.11
  [4536629a] OpenBLAS_jll v0.3.21+4
  [05823500] OpenLibm_jll v0.8.1+0
  [bea87d4a] SuiteSparse_jll v5.10.1+6
  [83775a58] Zlib_jll v1.2.13+0
  [8e850b90] libblastrampoline_jll v5.7.0+0
  [8e850ede] nghttp2_jll v1.48.0+0
  [3f19e933] p7zip_jll v17.4.0+0
Info Packages marked with ⌅ have new versions available but compatibility constraints restrict them from upgrading. To see why use `status --outdated -m`

You can also download the manifest file and the project file.