SymbolicNumericIntegration.jl

Documentation for SymbolicNumericIntegration.jl

SymbolicNumericIntegration.collect_powers — Method

collect_powers separates the powers of x in eq (a polynomial) and returns
a dictionary of power => term

SymbolicNumericIntegration.deg — Method

deg(p) returns the degree of p if p is a polynomial

SymbolicNumericIntegration.find_independent_subset — Method

returns a list of the indices of a linearly independent subset of the columns of A

SymbolicNumericIntegration.integrate — Function

integrate is the main entry point

input:
------
eq: a Symbolics expression to integrate
x: the independent variable (optional)

abstol: the desired tolerance
num_steps: the number of different steps with expanding basis to be tried
num_trials: the number of trials in each step (no changes to the basis)
radius: the radius of the disk in the complex plane to generate random test points
show_basis: if true, the basis (list of candidate terms) is printed
opt: the sparse regression optimizer
bypass: if true, do not integrate terms separately but consider all at once
symbolic: try symbolic integration first
max_basis: the maximum number of candidate terms to consider
verbose: print a detailed report
complex_plane: generate random test points on the complex plane (if false, the points will be on real axis)
homotomy: use the homotopy algorithm to generat the basis

output:
-------
solved, unsolved, err

solved is the solved integral and unsolved is the residual unsolved portion of the input
err is the numerical error in reaching the solution

SymbolicNumericIntegration.integrate_sum — Method

integrate_sum applies the integral summation rule ∫ Σᵢ fᵢ(x) dx = Σᵢ ∫ fᵢ(x) dx

eq: the integrand
x: the indepedent variable
l: a logger

output is the same as integrate

SymbolicNumericIntegration.integrate_term — Method

integrate_term is the central part of the code that tries different
methods to integrate eq

eq: the integrand
x: the indepedent variable
l: a logger

output is the same as integrate

SymbolicNumericIntegration.is_number — Method

is_number(x) returns true if x is a concrete numerical type

SymbolicNumericIntegration.isdependent — Method

isdependent returns true if eq is dependent on x

SymbolicNumericIntegration.nice_parameters — Method

converts float to int or small rational numbers

SymbolicNumericIntegration.solve_newton — Method

solve_newton is a symbolic Newton-Ralphson solver
f is a symbolic equation to be solved (f ~ 0)
x is the variable to solve
x₀ is the initial guess

SymbolicNumericIntegration.try_integrate — Method

the core of the randomized parameter-fitting algorithm

`try_integrate` tries to find a linear combination of the basis, whose
derivative is equal to eq

output
-------
integral, error

SymbolicNumericIntegration.var — Method

var(p) returns the unique variable of an expression (is exists)